+++ libs/Message.py
@@ -7,9 +7,9 @@ import json
 class Message(db.Model):
     __tablename__ = 'messages'
 
-    id = db.Column(db.BIGINT, nullable=False, primary_key=True, autoincrement=True)
-    chat_id = db.Column(db.Integer, db.ForeignKey('chats.id'), nullable=False)
-    user_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
+    id = db.Column(db.BigInteger, nullable=False, primary_key=True, autoincrement=True)
+    chat_id = db.Column(db.BigInteger, db.ForeignKey('chats.id'), nullable=False)
+    user_id = db.Column(db.BigInteger, db.ForeignKey('users.id'), nullable=False)
     time = db.Column(db.TIMESTAMP, default=timestamp())
     is_read = db.Column(db.Boolean, default=False)
     content = db.Column(db.JSON, default=json.dumps({'photos': [], 'audios': [], 'videos': [], 'map_pins': []}))
