+++ libs/Notification.py
@@ -4,8 +4,8 @@ from globals import db, timestamp
 class Notification(db.Model):
     __tablename__ = 'notifications'
 
-    user_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
-    chat_id = db.Column(db.Integer, db.ForeignKey('chats.id'), nullable=False)
+    user_id = db.Column(db.BigInteger, db.ForeignKey('users.id'), nullable=False)
+    chat_id = db.Column(db.BigInteger, db.ForeignKey('chats.id'), nullable=False)
     time = db.Column(db.TIMESTAMP, default=timestamp())
 
     __table_args__ = (
